TClock Light kt071223
* support for Vista's calendar.
* merge tclocklight-040702-3-amd64
  http://www.program-lab.jp/blog/progmemo/archives/2005/05/tclock_lightox6.html
* merge some patches from tclocklight-050716.

+++ source/sntp/dialog.c
@@ -178,7 +178,7 @@ void LoadLog(HWND hDlg)
 	add_title(fname, SNTPLOG);
 	hf = _lopen(fname, OF_READ);
 	if(hf == HFILE_ERROR) return;
-	_llseek(hf, -readlen, 2);
+	_llseek(hf, -(LONG)readlen, 2);
 	len = _lread(hf, buf, readlen);
 	_lclose(hf);
 	if(len == (UINT) -1) return;
